backend/common/requirements_v1.txt
aiofiles==0.7.0
python-multipart==0.0.5
openpyxl==3.0.9
python-magic==0.4.24